Modification related to recent branch re-base.
The code uses qint64 where it is possible for identifying messages. Also
removed some unused code.

+++ src/gui/dlg_msg_search.h
@@ -49,7 +49,7 @@ private slots:
 	void getSelectedMsg(int row, int column);
 
 signals:
-	void focusSelectedMsg(QString, int);
+	void focusSelectedMsg(QString, qint64);
 
 private:
 	const QList< QPair <QString,MessageDb*> > m_messageDbList;
